// src/app/models/address.ts
export interface Address {
  cep: string;
  address_type: string;
  address_name: string;
  address: string;
  district: string;
  state: string;
  city: string;
  lat: string;
  lng: string;
  ddd: string;
  city_ibge: string;
}
/*Interface criada inicialmente para servir de template (modelo) para os dados retornados pela API*/